requirements.txt
colorlog
# loguru
pyyaml
simplejson

cryptography
pycryptodome
